Moved `Term.Constant` and `Term.Variable` to the upper level.
