+++ lib/src/main/java/com/wjduquette/joe/nero/Constraint.java
@@ -1,6 +1,6 @@
 package com.wjduquette.joe.nero;
 
-public record Constraint(Term.Variable a, Op op, Term b) {
+public record Constraint(Variable a, Op op, Term b) {
     @Override public String toString() {
         return a + " " + op.lexeme() + " " + b;
     }
